# docs/swagger.yaml
basePath: /
definitions:
  main.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  main.Subscription:
    properties:
      end_date:
        type: string
      id:
        type: integer
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  main.SummaryResponse:
    properties:
      cost:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: completed test task for junior go developer
  title: TZ
  version: "0.1"
paths:
  /subscription:
    get:
      description: List all subscriptions
      operationId: list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Subscription'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: List subscriptions
    post:
      consumes:
      - application/json
      description: Create subscription
      operationId: create
      parameters:
      - description: Create data
        in: body
        name: createData
        required: true
        schema:
          $ref: '#/definitions/main.Subscription'
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Create Subscription
    put:
      consumes:
      - application/json
      description: Update subscription data
      operationId: update
      parameters:
      - description: Update data
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/main.Subscription'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Update Subscription
  /subscription/{id}:
    delete:
      description: Delete subscription by id
      operationId: delete
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Delete subscription
    get:
      description: Get subscription by id
      operationId: get
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Subscription'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Get subscription
  /subscription/summary:
    get:
      description: Get total cost of subscriptions for given filters
      operationId: cost
      parameters:
      - description: Start Date
        in: query
        name: start_date
        type: string
      - description: End Date
        in: query
        name: end_date
        type: string
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Service Name
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Get subscriptions cost
swagger: "2.0"